Issue #277 - More renames
