+++ plugin/src/main/java/org/jenkinsci/plugins/casc/impl/configurators/EnumConfigurator.java
@@ -1,7 +1,12 @@
-package org.jenkinsci.plugins.casc;
+package org.jenkinsci.plugins.casc.impl.configurators;
 
+import org.jenkinsci.plugins.casc.Attribute;
+import org.jenkinsci.plugins.casc.Configurator;
+import org.jenkinsci.plugins.casc.ConfiguratorException;
 import org.jenkinsci.plugins.casc.model.CNode;
 import org.jenkinsci.plugins.casc.model.Scalar;
+import org.kohsuke.accmod.Restricted;
+import org.kohsuke.accmod.restrictions.Beta;
 
 import javax.annotation.CheckForNull;
 import javax.annotation.Nonnull;
@@ -11,6 +16,7 @@ import java.util.Set;
 /**
  * @author <a href="mailto:nicolas.deloof@gmail.com">Nicolas De Loof</a>
  */
+@Restricted(Beta.class)
 public class EnumConfigurator<T extends Enum<T>> extends Configurator<T> {
 
     private final Class<T> clazz;
